deleted unnecessary files from older code, changed file locations for server and changed name of query file to resolver.
