+++ index.js
@@ -1,6 +1,6 @@
 require("dotenv").config();
 
-const server = require("./api/server");
+const server = require("./server");
 
 const port = process.env.PORT || 4000;
 server.listen(port, () => {
