+++ models/infodemand-model.js
@@ -5,12 +5,6 @@ const add = async function(info_row) {
   return insertedData;
 };
 
-const addition = async function(info_row) {
-  x = await db("info_demand_2").insert(info_row);
-  return x;
-};
-
 module.exports = {
-  add,
-  addition
+  add
 };
