+++ models/traders-model.js
