--- old code/seeds/003_commodity_seed.js
@@ -1,9 +1,0 @@
-exports.seed = function(knex, Promise) {
-  // Deletes ALL existing entries
-  return knex("commodity")
-    .del()
-    .then(function() {
-      // Inserts seed entries
-      return knex("commodity").insert([{ id: 1, name: "Maize" }]);
-    });
-};
