--- old code/tab3-routes/destMarket-model.js
@@ -1,11 +1,0 @@
-const db = require('../data/dbConfig');
-
-module.exports = {
-    getDestMarket
-};
-
-// Final Destination Market
-
-function getDestMarket() {
-    return db.raw('select * from information_demand where request_type_id = 8')
-};
