--- old code/tab3-routes/exchangeRate-model.js
@@ -1,10 +1,0 @@
-const db = require('../data/dbConfig');
-
-module.exports = {
-    getExRate
-};
-
-//Exchange Rate direction 
-function getExRate(){
-    return db('information_demand').where({request_type_id: '11'});
-}
