--- old code/tab3-routes/topCat-model.js
@@ -1,9 +1,0 @@
-const db = require('../data/dbConfig');
-
-module.exports = {
-    getTopCat
-};
-
-function getTopCat() {
-    return db('information_demand').where({request_type_id: '9'})
-}
