--- old code/tab3-routes/topCom-model.js
@@ -1,9 +1,0 @@
-const db = require('../data/dbConfig');
-
-module.exports = {
-    getTopCom
-};
-
-function getTopCom() {
-    return db('information_demand').where({request_type_id: '10'})
-}
